Create Documentation for Configuring Custom AMI
For this project we need quite a bit of software installed and configured.
To make this easier we have created a custom AMI to be used by Terraform in
the actual deployments.

This configuration will allow us to use Kafka as a systemd service, elasticsearch,
kibana, and Python. See setup.md for further details.